requirements.txt
Flask == 3.0.2
pymongo == 4.6.2
python-dotenv == 1.0.1
Authlib == 1.3.0
requests == 2.31.0
itsdangerous == 2.1.2
Flask-WTF == 1.2.1
email-validator == 2.1.1
Flask-Bcrypt == 1.0.1
bleach == 6.1.0
textstat == 0.7.3
language_tool_python == 2.7.1